Rename files from pascal case to snake case
