+++ python/app/widgets/collapse.py
